Introduce the GleanTestRule for the Glean SDK
This adds a test rule that can be added to encapsulate
the state of the Glean SDK in unit/instrumentation tests:
this rule takes care of resetting the Glean SDK state.
